iFeral / qb config detection, zsh, bash

+++ i
@@ -3,8 +3,8 @@
 # https://github.com/Aniverse/iFeral
 #
 #
-iFeralVer=0.3.9
-iFeralDate=2018.04.13.1
+iFeralVer=0.4.0
+iFeralDate=2018.04.14.1
 # 颜色 -----------------------------------------------------------------------------------
 black=$(tput setaf 0); red=$(tput setaf 1); green=$(tput setaf 2); yellow=$(tput setaf 3);
 blue=$(tput setaf 4); magenta=$(tput setaf 5); cyan=$(tput setaf 6); white=$(tput setaf 7);
@@ -94,43 +94,6 @@ echo -e "${bold}Ver. $iFeralDate    \n"
 
 mkdir -p ~/bin ~/lib ~/iFeral/backup ~/iFeral/log
 
-# 备份下，然后直接覆盖掉原先的内容
-cp -f ~/.profile ~/iFeral/backup/.profile."$(date "+%Y.%m.%d.%H.%M.%S")".bak >/dev/null 2>&1
-cat > ~/.profile <<EOF
-export LC_ALL=en_US.UTF-8
-export LANG=en_US.UTF-8
-export TZ="/usr/share/zoneinfo/Asia/Shanghai"
-
-export PATH=~/FH:~/iFeral/qb:~/iFeral/app:~/bin:~/pip/bin:~/.local/bin:\$PATH
-export LD_LIBRARY_PATH=~/lib:~/FH/link:\$LD_LIBRARY_PATH
-
-cdk=\$(df -h | grep `pwd | awk -F '/' '{print \$3}'` | awk '{print \$1}' | awk -F '/' '{print \$3}')
-[[ \$(echo \$cdk | grep -E "sd[a-z]+1") ]] && cdk=\$(echo \$cdk | sed "s/1//")
-alias io='iostat -d -x -m 1 | grep -E "\$cdk | rMB/s | wMB/s"'
-
-alias killde='kill "\$(pgrep -fu "\$(whoami)" "deluged")"'
-alias killde2='kill "\$(pgrep -fu "\$(whoami)" "de2")"'
-alias killtr='kill "\$(pgrep -fu "\$(whoami)" "transmission-daemon")"'
-alias killrt='kill "\$(pgrep -fu "\$(whoami)" "/usr/local/bin/rtorrent")"'
-alias killqb='kill \$(pgrep -fu "\$(whoami)" "qbitt")'
-
-alias cesu='echo;python ~/iFeral/app/spdtest --share;echo'
-alias cesu2='python ~/iFeral/app/spdtest --share --server'
-alias cesu3="echo;python ~/iFeral/app/spdtest --list 2>&1 | head -n30 | grep --color=always -P '(\d+)\.(\d+)\skm|(\d+)(?=\))';echo"
-alias ls="ls -hAv --color --group-directories-first"
-alias ll="ls -hAlvZ --color --group-directories-first"
-alias shanchu='rm -rf'
-alias zjpid='ps aux | egrep "$(whoami)|COMMAND" | egrep -v "grep|aux|root"'
-alias pid="ps aux | grep -v grep | grep"
-alias ios="iostat -d -x -m 1"
-alias wangsu='sar -n DEV 1| grep -E "rxkB\/s|txkB\/s|eth0|eth1"'
-alias scrgd="screen -R gooooogle"
-alias scrl="screen -ls"
-alias quanxian="chmod -R +x"
-alias cdb="cd .."
-alias gclone="git clone --depth=1"
-EOF
-
 fi ; }
 
 
@@ -148,9 +111,11 @@ echo -e "${green}(02) ${jiacu}安装 Deluge          "
 echo -e "${green}(04) ${jiacu}降级 rTorrent        "
 echo -e "${green}(05) ${jiacu}配置 ruTorrent       "
 echo -e "${green}(06) ${jiacu}安装 flexget         "
-echo -e "${green}(07) ${jiacu}安装 ffmpeg, rclone  "
+echo -e "${green}(07) ${jiacu}安装 ffmpeg          "
 echo -e "${green}(08) ${jiacu}查看 系统信息        "
 echo -e "${green}(09) ${jiacu}查看 邻居            "
+echo -e "${green}(10) ${jiacu}设置 .profile        "
+echo -e "${green}(11) ${jiacu}使用 zsh             "
 echo -e "${green}(99) ${jiacu}退出脚本             "
 echo -e "${normal}"
 
@@ -172,8 +137,7 @@ case $response in
     7 | 07) # 安装 一些工具
             _install_tools ;;
     8 | 08) # 查看 系统信息
-            _stats
-            _main_menu ;;
+            _stats ; _main_menu ;;
     9 | 09) # 查看 所有邻居
             clear
             echo -e "${bold}${cayn}以下是当前和你在同一个硬盘分区上的邻居${normal}\n"
@@ -181,6 +145,7 @@ case $response in
             echo -e "${bold}${cayn}以下是整个盒子上所有的邻居${normal}\n"
             getent passwd | grep -v $(whoami) | grep -E 'home[0-9]+|media' | awk -F ':' '{print $1}' | sort -u | pr -3 -t ; echo
             _main_menu ;;
+        10) _set_profile ; _main_menu ;;
     99| "") clear ; exit 0 ;;
     *     ) clear ; exit 0 ;;
 esac
@@ -195,12 +160,12 @@ echo ; }
 
 function _install_qb() {
 
-echo
+echo ; mkdir -p ~/tmp ~/private/qbittorrent/{data,watch,torrents} ~/.config/{qBittorrent,flexget}
 
-mkdir -p ~/tmp ~/private/qbittorrent/{data,watch,torrents} ~/.config/{qBittorrent,flexget}
-
+# 关掉可能在运行的 qb
 for qbpid in ` ps aux | grep $(whoami) | grep -Ev "grep|aux|root" | grep qbittorrent | awk '{print $2}' ` ; do kill -9 $qbpid ; done
 
+# 下载
 if [[ ! `  ls ~/iFeral/qb/library  2>/dev/null  `  ]]; then
     echo -e "${bold}${yellow}下载 qbittorrent-nox ...${normal}\n"
     if   [[ $Seedbox == FH ]]; then
@@ -213,22 +178,35 @@ if [[ ! `  ls ~/iFeral/qb/library  2>/dev/null  `  ]]; then
     chmod +x -R ~/iFeral/qb ; echo
 fi
 
+# 询问版本
 while [[ $QBVERSION = "" ]]; do
     [[ $Seedbox == SH ]] && echo -e "${atte} SH 没有 4.0 及以后的版本可选！"
     echo -ne "${bold}${yellow}请输入你要安装的 qBittorrent 版本，只支持 3.3.0-4.0.4: ${normal}" ; read -e QBVERSION
-    while [[ ! $QBVERSION = "" ]]; do
+    while [[ $QBVERSION = "" ]]; do
         [[ ! ` ls ~/iFeral/qb | grep $QBVERSION ` ]] && { echo -e "${error} 你输入的版本不可用，请重新输入！" ; unset QBVERSION ; }
     done
 done
 
-read -ep "${bold}${yellow}请输入你要用于 qb WebUI 的密码：${normal}" PASSWORD
-QBPASS=`  echo -n $PASSWORD | md5sum | awk '{print $1}'  `
+# 询问是否覆盖原配置信息
+if [[ -e ~/.config/qBittorrent/qBittorrent.conf ]]; then
+    echo -e "${atte} 你以前装过 qBittorrent，那时的配置文件还留着，包含着 qBittorrent 的账号、密码、端口、下载路径等信息"
+    echo -ne "你现在要使用以前留下的配置文件吗？${normal} [${cyan}Y${normal}]es or [N]o: " ; read -e responce
+    case $responce in
+        [yY] | [yY][Ee][Ss] | "" ) qbconfig=old ;;
+        [nN] | [nN][Oo]          ) qbconfig=new ;;
+        *                        ) qbconfig=old ;;
+    esac
+else
+    qbconfig=new
+fi
 
-portGenerator && portCheck
-portGenerator2 && portCheck2
-
-    if [[ ! -e ~/.config/qBittorrent/qBittorrent.conf ]]; then
-cat > ~/.config/qBittorrent/qBittorrent.conf <<EOF
+# 新建配置文件
+if [[ $qbconfig == new ]]; then
+    read -ep "${bold}${yellow}请输入你要用于 qb WebUI 的密码：${normal}" PASSWORD
+    QBPASS=`  echo -n $PASSWORD | md5sum | awk '{print $1}'  `
+    portGenerator && portCheck
+    portGenerator2 && portCheck2
+    cat > ~/.config/qBittorrent/qBittorrent.conf <<EOF
 [Application]
 FileLogger\Enabled=true
 FileLogger\Age=6
@@ -263,22 +241,27 @@ WebUI\Port=$portGen
 WebUI\Password_ha1=@ByteArray($QBPASS)
 WebUI\Username=$(whoami)
 EOF
-    fi
+fi
 
 QBPORT=` grep "WebUI.Port" ~/.config/qBittorrent/qBittorrent.conf | grep -Po "\d+" `
+QBUSERNAME=` grep "WebUI.Username" ~/.config/qBittorrent/qBittorrent.conf | awk -F "=" '{print $NF}' `
 
+# 运行 qBittorrent-nox
 TMPDIR=~/tmp LD_LIBRARY_PATH=~/iFeral/qb/library ~/iFeral/qb/qbittorrent-nox.$QBVERSION -d
 
+# 输出结果
 if [[ ` ps aux | grep $(whoami) | grep -Ev "grep|aux|root" | grep qbittorrent ` ]]; then
-    echo -e "${bold}${green}
-qBittorrent 已安装完成！${jiacu}
-
-网址  ${cyan}http://$(hostname -f):$QBPORT${jiacu}
-账号  ${cyan}$(whoami)${jiacu}
-密码  ${cyan}$PASSWORD${normal}
-注意  如果你原先装过一次 qb，那你应该使用之前的账号密码登陆 qb"
+    echo -e "\n${bold}${green}"
+    echo -e "qBittorrent 已安装完成！${jiacu}\n"
+    echo -e "网址  ${cyan}http://$(hostname -f):$QBPORT${jiacu}"
+    [[ $qbconfig == new ]] &&
+    echo -e "账号  ${cyan}$QBUSERNAME${jiacu}" &&
+    echo -e "密码  ${cyan}$PASSWORD${normal}"
+    [[ $qbconfig == old ]] &&
+    echo -e "账号  ${cyan}$QBUSERNAME${jiacu}" &&
+    echo -e "密码  ${cyan}(和以前一样)${normal}"
 else
-    echo -e "${error} qBittorrent 安装完成，但无法正常运行。\n不要问我为什么，我可能也不知道！要不你换个别的脚本试试？${normal}"
+    echo -e "${error} qBittorrent 安装完成，但无法正常运行。\n不要问我为什么和怎么办，你自己看着办吧！${normal}"
 fi ; }
 
 
@@ -291,6 +274,8 @@ fi ; }
 function _install_de() {
 
 echo ; cd
+
+# 关掉可能在运行的第二个 deluged
 for depid in ` ps aux | grep $(whoami) | grep -Ev "grep|aux|root" | grep de2 | awk '{print $2}' ` ; do kill -9 $depid ; done
 
 while [[ $DEVERSION = "" ]]; do
@@ -340,7 +325,7 @@ daemon  账号  ${cyan}$(whoami)${jiacu}
 daemon  密码  ${cyan}$(sed -rn "s/$(whoami):(.*):(.*)/\1/p" ~/.config/deluge2/auth)${jiacu}
 daemon  端口  ${cyan}$portGen${normal}"
 else
-    echo -e "${error} 第二个 Deluge 安装完成，但无法正常运行。\n不要问我为什么，我可能也不知道！要不你手动安装试试？${normal}"
+    echo -e "${error} 第二个 Deluge 安装完成，但无法正常运行。\n不要问我为什么和怎么办，你自己看着办吧！${normal}"
 fi ; }
 
 
@@ -507,18 +492,18 @@ make -j$(nproc) && make install DEST_HOME=$HOME
 cd && rm -f ~/p7zip.tar.bz2
 
 # rclone
-mkdir -p ~/bin
-wget -qO ~/rclone.zip http://downloads.rclone.org/rclone-current-linux-amd64.zip
-unzip -qq ~/rclone.zip
-mv ~/rclone-v*-linux-amd64/rclone ~/bin
-rm -rf ~/rclone-v*-linux-amd64 ~/rclone.zip
-chmod +x ~/bin/rclone
+# mkdir -p ~/bin
+# wget -qO ~/rclone.zip http://downloads.rclone.org/rclone-current-linux-amd64.zip
+# unzip -qq ~/rclone.zip
+# mv ~/rclone-v*-linux-amd64/rclone ~/bin
+# rm -rf ~/rclone-v*-linux-amd64 ~/rclone.zip
+# chmod +x ~/bin/rclone
 
 # mktorrent
-git clone --depth=1 https://github.com/Rudde/mktorrent
-cd mktorrent/ && PREFIX=$HOME make -j$(nproc)
-PREFIX=$HOME make install
-cd .. && rm -rf mktorrent
+# git clone --depth=1 https://github.com/Rudde/mktorrent
+# cd mktorrent/ && PREFIX=$HOME make -j$(nproc)
+# PREFIX=$HOME make install
+# cd .. && rm -rf mktorrent
 
 }
 
@@ -618,6 +603,145 @@ echo -e  "${bold}  拥塞控制算法  : ${cyan}$tcp_control${normal}"
 echo
 echo -e  "${bold}  服务器时间    : ${cyan}$date${normal}"
 echo ; }
+
+
+
+
+
+
+
+
+
+
+# 10. 设置 bash 环境
+function _set_profile() { 
+
+# 备份下，然后直接覆盖掉原先的内容
+cp -f ~/.profile ~/iFeral/backup/.profile."$(date "+%Y.%m.%d.%H.%M.%S")".bak >/dev/null 2>&1
+cat > ~/.profile <<EOF
+export LC_ALL=en_US.UTF-8
+export LANG=en_US.UTF-8
+export TZ="/usr/share/zoneinfo/Asia/Shanghai"
+
+export PATH=~/FH:~/iFeral/qb:~/iFeral/app:~/bin:~/pip/bin:~/.local/bin:\$PATH
+
+cdk=\$(df -h | grep `pwd | awk -F '/' '{print \$(NF-1)}'` | awk '{print \$1}' | awk -F '/' '{print \$3}')
+[[ \$(echo \$cdk | grep -E "sd[a-z]+1") ]] && cdk=\$(echo \$cdk | sed "s/1//")
+alias io='iostat -d -x -m 1 | grep -E "\$cdk | rMB/s | wMB/s"'
+
+alias killde='kill "\$(pgrep -fu "\$(whoami)" "deluged")"'
+alias killde2='kill "\$(pgrep -fu "\$(whoami)" "de2")"'
+alias killtr='kill "\$(pgrep -fu "\$(whoami)" "transmission-daemon")"'
+alias killrt='kill "\$(pgrep -fu "\$(whoami)" "/usr/local/bin/rtorrent")"'
+alias killqb='kill \$(pgrep -fu "\$(whoami)" "qbitt")'
+
+alias cesu='echo;python ~/iFeral/app/speedtest --share;echo'
+alias cesu2='python ~/iFeral/app/speedtest --share --server'
+alias cesu3="echo;python ~/iFeral/app/speedtest --list 2>&1 | head -n30 | grep --color=always -P '(\d+)\.(\d+)\skm|(\d+)(?=\))';echo"
+alias ls="ls -hAv --color --group-directories-first"
+alias ll="ls -hAlvZ --color --group-directories-first"
+alias shanchu='rm -rf'
+alias zjpid='ps aux | egrep "$(whoami)|COMMAND" | egrep -v "grep|aux|root"'
+alias pid="ps aux | grep -v grep | grep"
+alias ios="iostat -d -x -m 1"
+alias wangsu='sar -n DEV 1| grep -E "rxkB\/s|txkB\/s|eth0|eth1"'
+alias scrgd="screen -R gooooogle"
+alias scrl="screen -ls"
+alias quanxian="chmod -R +x"
+alias cdb="cd .."
+alias gclone="git clone --depth=1"
+EOF
+}
+
+
+
+
+
+# 11. 设置 zsh 环境
+function _set_zsh() { if [[ `command -v zsh` ]]; then
+cd ; echo -e "\n${atte} 等会儿切换到 zsh 需要输入当前 SSH 的密码${normal}\n"
+rm -rf ~/.oh-my-zsh .zshrc
+sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+wget -qO ~/.oh-my-zsh/themes/agnosterzak.zsh-theme http://raw.github.com/zakaziko99/agnosterzak-ohmyzsh-theme/master/agnosterzak.zsh-theme
+sed -i "s/robbyrussell/agnosterzak/g" ~/.zshrc
+
+cat >> ~/.zshrc <<EOF
+export LC_ALL=en_US.UTF-8
+export LANG=en_US.UTF-8
+export TZ="/usr/share/zoneinfo/Asia/Shanghai"
+
+export PATH=~/FH:~/iFeral/qb:~/iFeral/app:~/bin:~/pip/bin:~/.local/bin:\$PATH
+
+alias -s sh='bash'
+alias -s log='tail -n50'
+alias -s gz='tar zxf'
+alias -s xz='tar xf'
+alias -s tgz='tar zxf'
+alias -s rar='unrar x'
+alias -s zip='unzip'
+alias -s bz2='tar -xjvf'
+
+cdk=\$(df -h | grep `pwd | awk -F '/' '{print \$(NF-1)}'` | awk '{print \$1}' | awk -F '/' '{print \$3}')
+[[ \$(echo \$cdk | grep -E "sd[a-z]+1") ]] && cdk=\$(echo \$cdk | sed "s/1//")
+alias io='iostat -d -x -m 1 | grep -E "\$cdk | rMB/s | wMB/s"'
+
+alias killde='kill "\$(pgrep -fu "\$(whoami)" "deluged")"'
+alias killde2='kill "\$(pgrep -fu "\$(whoami)" "de2")"'
+alias killtr='kill "\$(pgrep -fu "\$(whoami)" "transmission-daemon")"'
+alias killrt='kill "\$(pgrep -fu "\$(whoami)" "/usr/local/bin/rtorrent")"'
+alias killqb='kill \$(pgrep -fu "\$(whoami)" "qbitt")'
+
+alias cesu='echo;python ~/iFeral/app/speedtest --share;echo'
+alias cesu2='python ~/iFeral/app/speedtest --share --server'
+alias cesu3="echo;python ~/iFeral/app/speedtest --list 2>&1 | head -n30 | grep --color=always -P '(\d+)\.(\d+)\skm|(\d+)(?=\))';echo"
+alias ls="ls -hAv --color --group-directories-first"
+alias ll="ls -hAlvZ --color --group-directories-first"
+alias shanchu='rm -rf'
+alias zjpid='ps aux | egrep "$(whoami)|COMMAND" | egrep -v "grep|aux|root"'
+alias pid="ps aux | grep -v grep | grep"
+alias ios="iostat -d -x -m 1"
+alias wangsu='sar -n DEV 1| grep -E "rxkB\/s|txkB\/s|eth0|eth1"'
+alias scrgd="screen -R gooooogle"
+alias scrl="screen -ls"
+alias quanxian="chmod -R +x"
+alias cdb="cd .."
+alias gclone="git clone --depth=1"
+
+# Fix numeric keypad  
+# 0 . Enter  
+bindkey -s "^[Op" "0"  
+bindkey -s "^[On" "."  
+bindkey -s "^[OM" "^M"  
+# 1 2 3  
+bindkey -s "^[Oq" "1"  
+bindkey -s "^[Or" "2"  
+bindkey -s "^[Os" "3"  
+# 4 5 6  
+bindkey -s "^[Ot" "4"  
+bindkey -s "^[Ou" "5"  
+bindkey -s "^[Ov" "6"  
+# 7 8 9  
+bindkey -s "^[Ow" "7"  
+bindkey -s "^[Ox" "8"  
+bindkey -s "^[Oy" "9"  
+# + - * /  
+bindkey -s "^[Ol" "+"  
+bindkey -s "^[Om" "-"  
+bindkey -s "^[Oj" "*"  
+bindkey -s "^[Oo" "/"  
+EOF
+
+chsh -s /usr/bin/zsh
+source ~/.zshrc
+
+else echo -e "\n${error} 你的盒子没有预装 zsh，故无法使用本功能！${normal}\n" ; fi ; }
+
+
+
+
+
+
+
